refactor: Add custom hook for navigating to user profile screen
