+++ src/components/organisms/Community/CommentListItem.tsx
@@ -1,11 +1,10 @@
 import { View, StyleSheet } from 'react-native';
-import React, { useCallback } from 'react';
+import React from 'react';
 import CommentContent from '/components/atoms/Community/CommentContent';
 import CommentHeader from '/components/molecules/Community/CommentHeader';
 import CommentFooter from '/components/molecules/Community/CommentFooter';
 import CustomAvatar from '/components/atoms/Common/CustomAvatar';
-import { useNavigation } from '@react-navigation/native';
-import { useAuthValue } from '/hooks/context/useAuth';
+import { useNavigateToUser } from '/hooks/common/navToUserProfile';
 
 type Props = {
   comment: CommentRead;
@@ -16,19 +15,7 @@ export default function CommentListItem({
   comment,
   onPressEditComment,
 }: Props) {
-  const nav = useNavigation();
-  const authInfo = useAuthValue();
-  const navigateToUser = useCallback(() => {
-    nav.navigate('MainNavigator', {
-      screen: '커뮤니티',
-      params: {
-        screen: 'User',
-        params: {
-          userId: comment.user.id === authInfo?.userId ? 'me' : comment.user.id,
-        },
-      },
-    });
-  }, [nav, comment.user.id, authInfo?.userId]);
+  const navigateToUser = useNavigateToUser(comment.user.id);
 
   return (
     <View style={styles.container}>
