+++ src/components/organisms/Community/PostDetailSection.tsx
@@ -8,8 +8,7 @@ import { Button, Divider, Text } from 'react-native-paper';
 import { ErrorBoundary, FallbackProps } from 'react-error-boundary';
 import { usePostQuery } from '/hooks/queries/post.queries';
 import { useQueryErrorResetBoundary } from '@tanstack/react-query';
-import { useNavigation } from '@react-navigation/native';
-import { useAuthValue } from '/hooks/context/useAuth';
+import { useNavigateToUser } from '/hooks/common/navToUserProfile';
 
 type Props = {
   postId: number;
@@ -24,21 +23,8 @@ export default function PostDetailSection({
 }: Props) {
   const { data: post, refetch } = usePostQuery(postId);
   const { reset } = useQueryErrorResetBoundary();
-  const nav = useNavigation();
-  const authInfo = useAuthValue();
-  const navigateToUser = useCallback(() => {
-    if (!post) return;
 
-    nav.navigate('MainNavigator', {
-      screen: '커뮤니티',
-      params: {
-        screen: 'User',
-        params: {
-          userId: post.user.id === authInfo?.userId ? 'me' : post.user.id,
-        },
-      },
-    });
-  }, [nav, post, authInfo?.userId]);
+  const navigateToUser = useNavigateToUser(post?.user.id);
 
   const renderPostError = useCallback(
     ({ error, resetErrorBoundary }: FallbackProps) => {
